Move array manipulation function in xmanipulation
- Also implement rot90 function
- And fix minor issues with data offset for flip and transpose
