+++ test/test_xarray.cpp
@@ -5,10 +5,11 @@
 *                                                                          *
 * The full license is in the file LICENSE, distributed with this software. *
 ****************************************************************************/
- 
+
 #include "gtest/gtest.h"
 #include "xtensor/xarray.hpp"
 #include "xtensor/xtensor.hpp"
+#include "xtensor/xmanipulation.hpp"
 #include "xtensor/xio.hpp"
 #include "test_common.hpp"
