+++ test/test_xview.cpp
@@ -13,6 +13,7 @@
 #include "xtensor/xfixed.hpp"
 #include "xtensor/xnoalias.hpp"
 #include "xtensor/xstrided_view.hpp"
+#include "xtensor/xmanipulation.hpp"
 #include "xtensor/xtensor.hpp"
 #include "xtensor/xview.hpp"
 #include "xtensor/xrandom.hpp"
